+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+- Fix `shape.getClientRect()` when any of parents is cached
+
 ### 9.3.11 (2024-05-23)
 
 - Fix chrome clear canvas issue